+++ megamek/src/megamek/common/MechBay.java
@@ -97,8 +97,8 @@ public final class MechBay extends Bay {
     }
 
     @Override
-    public float getWeight() {
-        return (float) totalSpace * 150;
+    public double getWeight() {
+        return totalSpace * 150;
     }
 
     @Override
